eppolicymap: make sure EpPolicyMap is created to avoid nil dereference
In eppolicymap.CreateWithName(), if the policy map creation in the
buildMap.Do() block fails for whatever reason, we return early and
EpPolicyMap remains nil. Calling OpenOrCreate() a few lines below
attempts to dereference that pointer to take its lock.

Ouch.

EpPolicyMap used to be created unconditionally, even if the policy map
creation would fail. This was changed in commit b384ce85bb83. Let's move
its creation above the policy map creation, so that we are sure we don't
leave with a nil pointer.

Fixes: b384ce85bb83 ("eppolicymap: Factor out for better testing")
Signed-off-by: Quentin Monnet <quentin@isovalent.com>