fragmap, policymap: clean-up map size configuration
As a follow-up to IPv4 fragmentation cleanup (#10927), clean-up
configuration for the maximum number of entries in policy map and
fragment-tracking map.

Remove unnecessary InitMapInfo() for fragmap. Keep it for policy map,
because passing the size at map creation and removing this function
would mean:

- Updating the list of arguments passed to Open(), OpenAndCreate(),
  Create() and pass the size (option.Config.PolicyMapEntries) each time,
  which seems verbose and not really useful. We cannot import the
  "option" package in "policy" to use the option just once in newMap().

- Making sure that policymap.MaxEntries is never called from other
  packages before the map is created. This is probably not the case so
  we do want to set the size before map creation.

Initialise policymap.MaxEntries to defaults.PolicyMapEntries.
Theoretically we don't have to initialise it because it is overwritten
by InitMapInfo() before use, but the unit tests do not call
InitMapInfo() and rely a non-zero value to successfully create a map
(eppolicymap unit tests also use the value but does not care about map
creation being successful).

Signed-off-by: Quentin Monnet <quentin@isovalent.com>